Fix precommit
